# Make extraction faster