--- a/tsfresh/feature_extraction/extraction.py
+++ b/tsfresh/feature_extraction/extraction.py
@@ -120,7 +120,8 @@ def extract_features(timeseries_container, feature_extraction_settings=None,
         result = _extract_features_parallel_per_sample(kind_to_df_map, feature_extraction_settings,
                                                        column_id, column_value)
     elif parallelization == 'no_parallelization':
-        result = _extract_features_serial(kind_to_df_map, feature_extraction_settings, column_id, column_value)
+        result = _extract_features_parallel_per_kind(kind_to_df_map, feature_extraction_settings,
+                                                     column_id, column_value, serial=True)
     else:
         raise ValueError("Argument parallelization must be one of: 'per_kind', 'per_sample'")
 
@@ -132,7 +133,7 @@ def extract_features(timeseries_container, feature_extraction_settings=None,
     return result
 
 
-def _extract_features_parallel_per_kind(kind_to_df_map, settings, column_id, column_value):
+def _extract_features_parallel_per_kind(kind_to_df_map, settings, column_id, column_value, serial=False):
     """
     Parallelize the feature extraction per kind.
 
@@ -147,6 +148,9 @@ def _extract_features_parallel_per_kind(kind_to_df_map, settings, column_id, col
     :param settings: settings object that controls which features are calculated
     :type settings: tsfresh.feature_extraction.settings.FeatureExtractionSettings
 
+    :param serial: Do not parallelize the extraction. This can be handy if (1) you want to debug something
+       (2) you want to profile something or (3) your environment does not support multiprocessing
+
     :return: The (maybe imputed) DataFrame containing extracted features.
     :rtype: pandas.DataFrame
     """
@@ -159,8 +163,14 @@ def _extract_features_parallel_per_kind(kind_to_df_map, settings, column_id, col
     chunksize = helper_functions.calculate_best_chunksize(kind_to_df_map, settings)
 
     total_number_of_expected_results = len(kind_to_df_map)
-    extracted_features = tqdm(pool.imap_unordered(partial_extract_features_for_one_time_series, kind_to_df_map.items(),
-                                                  chunksize=chunksize), total=total_number_of_expected_results,
+
+    if serial:
+        map_function = map
+    else:
+        map_function = partial(pool.imap_unordered, chunksize=chunksize)
+
+    extracted_features = tqdm(map_function(partial_extract_features_for_one_time_series, kind_to_df_map.items()),
+                              total=total_number_of_expected_results,
                               desc="Feature Extraction", disable=settings.disable_progressbar)
 
     pool.close()
@@ -174,45 +184,6 @@ def _extract_features_parallel_per_kind(kind_to_df_map, settings, column_id, col
 
     pool.join()
     return result
-
-
-def _extract_features_serial(kind_to_df_map, settings, column_id, column_value):
-    """
-    Feature extraction without parallelization.
-
-    :param kind_to_df_map: The time series to compute the features for in our internal format
-    :type kind_to_df_map: dict of pandas.DataFrame
-
-    :param column_id: The name of the id column to group by.
-    :type column_id: str
-    :param column_value: The name for the column keeping the value itself.
-    :type column_value: str
-
-    :param settings: settings object that controls which features are calculated
-    :type settings: tsfresh.feature_extraction.settings.FeatureExtractionSettings
-
-    :return: The (maybe imputed) DataFrame containing extracted features.
-    :rtype: pandas.DataFrame
-    """
-    partial_extract_features_for_one_time_series = partial(_extract_features_for_one_time_series,
-                                                           column_id=column_id,
-                                                           column_value=column_value,
-                                                           settings=settings)
-
-    total_number_of_expected_results = len(kind_to_df_map)
-    extracted_features = tqdm(map(partial_extract_features_for_one_time_series, kind_to_df_map.items()),
-                              total=total_number_of_expected_results,
-                              desc="Feature Extraction", disable=settings.disable_progressbar)
-
-    # Concatenate all partial results
-    result = pd.concat(extracted_features, axis=1, join='outer').astype(np.float64)
-
-    # Impute the result if requested
-    if settings.IMPUTE is not None:
-        settings.IMPUTE(result)
-
-    return result
-
 
 def _extract_features_parallel_per_sample(kind_to_df_map, settings, column_id, column_value):
     """