This one was some major work, so let me describe everything in detail:
I have done some profiling to increase the runtime performance of the feature extraction. 

I have done this using two tools (as reference for later): cProfile + pyprof2calltree for vizualisation and kernprof (from line_profiling), e.g. with

    python -m cProfile -o prof.out tests/examples/test_robot_execution_failures.py
    pyprof2calltree -i prof.out -k
    kernprof -l -v tests/examples/test_robot_execution_failures.py

Everything was done on our default robots failures dataset (I know that this may not be the perfect example for out usage, but I need to use something that is standardized).

After identifying the main runtime issues, I have refactored them step by step. I have made sure, that the results stayed the same during all steps using the full robots dataset and our unit tests.

Here are the results:

| What | Before | After |
| ------- | --------- | ------- |
| extraction on full robot set (without parallelization) | 39.51s | 20.68s ( - 48% !!!) |
| run all unittest | 41.91s | 38.73s | 

Parallelisation was turned off for the first measurement, to not rely on kernel process management etc.

The changes to the functions were in detail (all in s for the first 20 ids of the robot set):

| Name | Before | After |
| ---- | ---- | ---- |
| large_standard_dev | 0.33 | 0.076 |
| symmetry_looking | 0.758 | 0.046 |
| number_peaks | 3.901 | 0.160 |
| index_mass_quantile | 0.926 | 0.039 |
| cwt_coeff | 1.813 | 0.129 |
| ar_coeff | 0.626 | 0.1629 |

After that, we only have 4 functions left, that have "large" performance issues:
* augmented_dicker_fuller 14 %
* mean_abs_change_qunatiles 25 %
* autocorrelation 14 %
* quantile 6 %

The percentage numbers are the percentage they use up during the full extraction function.

TLDR: I increased the runtime by approx 50% on our test dataset while remaining the content of the calculations.