@MaxBenChrist 2017-04-02
> TLDR: I increased the runtime by approx 50% on our test dataset while remaining the content of the calculations.

Your mean Decreased, right?

 I will go over this Pr tomorrow. Looks like a new release is coming in ;)