@nils-braun 2017-04-05
> from what i saw, you used the following techniques to speed up calculations

> * cache results by making it apply from aggregate (symmetry_looking)
> * do not sort parameters, use dict instead (ar_coefficient, ctw coefficients)
> * ?? (number_peaks)

One of the major things for increasing the performance was the reduction of the usage of pandas dataframes and series. They are very handle, but unfortunately very slow - especially when it comes to appending or concationating of results. Here, it is much better to store all results in a python list/dict and create a Series/DataFrame when everything is finished. Also, applying mean/std etc. and indexing is much faster on pure numpy arrays than on pandas Series.
Another optimization was to not use DataFrames for the configurations (e.g. in cwt_coefficients), because we would need to groupby/index/mask many times which is slow. Better use a cache for the calculated values and work through all configuration parameters one-by-one.