@MaxBenChrist 2017-04-05
> Another optimization was to not use DataFrames for the configurations (e.g. in cwt_coefficients), because we would need to groupby/index/mask many times which is slow. Better use a cache for the calculated values and work through all configuration parameters one-by-one.

This was an aha-moment for me. Dictionaries are great.