@MaxBenChrist 2017-04-05
> And I do not know why the coverage has dropped...

Normally, you can just click on the details field, the coveralls page shows you the lines where coverage is missing. But here it says that it can't find the source file? strange