@MaxBenChrist 2017-04-06
So, the only thing that is missing is the adaption of the _extract_features_parallel_per_kind method to do it in serial. Then we can merge this thing here. Maybe this also helps with the coverage